+++ test/service_extensions_test.dart
@@ -1,68 +1,72 @@
 import 'dart:developer' as dev;
 
 import 'package:service_extensions/service_extensions.dart';
-import 'package:service_extensions/src/registery.dart'
-    show initDiscoveryRegistery;
+import 'package:service_extensions/src/registry.dart' show initDiscoveryRegistry;
 import 'package:test/test.dart';
 
 void main() {
-  group('ExtensionParameters', () {
+  group('ServiceExtensionParameters', () {
     test('asString', () {
-      final params = ExtensionParameters({'foo': 'bar'}, method: 'test');
+      final params =
+          ServiceExtensionParameters({'foo': 'bar'}, method: 'test');
       expect(params.asString('foo'), 'bar');
       expect(params.asString('baz'), isNull);
     });
 
     test('asStringRequired', () {
-      final params = ExtensionParameters({'foo': 'bar'}, method: 'test');
+      final params =
+          ServiceExtensionParameters({'foo': 'bar'}, method: 'test');
       expect(params.asStringRequired('foo'), 'bar');
       expect(() => params.asStringRequired('baz'),
           throwsA(isA<dev.ServiceExtensionResponse>()));
     });
 
     test('asBool', () {
-      final params =
-          ExtensionParameters({'foo': 'true', 'bar': 'false'}, method: 'test');
+      final params = ServiceExtensionParameters({'foo': 'true', 'bar': 'false'},
+          method: 'test');
       expect(params.asBool('foo'), isTrue);
       expect(params.asBool('bar'), isFalse);
       expect(params.asBool('baz'), isNull);
 
       final badParams =
-          ExtensionParameters({'foo': 'not-a-bool'}, method: 'test');
+          ServiceExtensionParameters({'foo': 'not-a-bool'}, method: 'test');
       expect(() => badParams.asBool('foo'),
           throwsA(isA<dev.ServiceExtensionResponse>()));
     });
 
     test('asBoolRequired', () {
-      final params = ExtensionParameters({'foo': 'true'}, method: 'test');
+      final params = ServiceExtensionParameters({'foo': 'true'}, method: 'test');
       expect(params.asBoolRequired('foo'), isTrue);
       expect(() => params.asBoolRequired('baz'),
           throwsA(isA<dev.ServiceExtensionResponse>()));
     });
 
     test('asInt', () {
-      final params = ExtensionParameters({'foo': '123'}, method: 'test');
+      final params = ServiceExtensionParameters({'foo': '123'}, method: 'test');
       expect(params.asInt('foo'), 123);
       expect(params.asInt('bar'), isNull);
 
-      final badParams = ExtensionParameters({'foo': 'abc'}, method: 'test');
+      final badParams =
+          ServiceExtensionParameters({'foo': 'abc'}, method: 'test');
       expect(() => badParams.asInt('foo'),
           throwsA(isA<dev.ServiceExtensionResponse>()));
     });
 
     test('asIntRequired', () {
-      final params = ExtensionParameters({'foo': '123'}, method: 'test');
+      final params = ServiceExtensionParameters({'foo': '123'}, method: 'test');
       expect(params.asIntRequired('foo'), 123);
       expect(() => params.asIntRequired('bar'),
           throwsA(isA<dev.ServiceExtensionResponse>()));
     });
 
     test('asDouble', () {
-      final params = ExtensionParameters({'foo': '123.45'}, method: 'test');
+      final params =
+          ServiceExtensionParameters({'foo': '123.45'}, method: 'test');
       expect(params.asDouble('foo'), 123.45);
       expect(params.asDouble('bar'), isNull);
 
-      final badParams = ExtensionParameters({'foo': 'abc'}, method: 'test');
+      final badParams =
+          ServiceExtensionParameters({'foo': 'abc'}, method: 'test');
       expect(() => badParams.asDouble('foo'),
           throwsA(isA<dev.ServiceExtensionResponse>()));
     });
@@ -70,7 +74,7 @@ void main() {
 
   group('ServiceDescription', () {
     test('tracking registered extensions', () {
-      initDiscoveryRegistery();
+      initDiscoveryRegistry();
 
       final listExtensions = registeredExtensions
           .firstWhere((e) => e.name == 'ext.service_extensions.getExtensions');
@@ -94,6 +98,31 @@ void main() {
       expect(json['type'], 'String');
       expect(json['description'], 'a bar');
       expect(json['required'], isTrue);
+    });
+
+    test('equality', () {
+      final d1 = ServiceDescription(
+        name: 'foo',
+        description: 'bar',
+        parameters: [
+          ParameterDescription(name: 'p1', type: 'int', description: 'p1 desc')
+        ],
+      );
+      final d2 = ServiceDescription(
+        name: 'foo',
+        description: 'bar',
+        parameters: [
+          ParameterDescription(name: 'p1', type: 'int', description: 'p1 desc')
+        ],
+      );
+      final d3 = ServiceDescription(
+        name: 'foo',
+        description: 'baz',
+      );
+
+      expect(d1, equals(d2));
+      expect(d1, isNot(equals(d3)));
+      expect(d1.hashCode, equals(d2.hashCode));
     });
   });
 }